Update to Rev1.1 files
In Rev 1.0, GND pin of “RESET” button(S1) was floating.
This change in Rev1.1 will solve the floating problem. (Connect to GND
with adding VIA.)
Refer to below picture.
GND is reinforced by adding via hole to whole board
Via located was close to D11 pattern(R20). so that via is separated.